refactor(activemodel,activerecord): split TS-only translation tests, converge Table option splats, dedupe MySQL create_table_definition
Three RFC 0115 / RFC 0051 convergence stories.

1. activemodel `translation.test.ts` mirrors `translation_test.rb` and now
   holds only its 26 Rails test names; the 19 TS-only cases (6 duplicate-name
   copies plus the dotted-attribute and `raise_on_missing_translations`
   groups) move to `translation.trails.test.ts`. Where a Rails name appeared
   twice, the copy kept in the mirrored file is the one whose body follows
   the Ruby (`translation_test.rb:121-140`).

2. Ruby's `**options` passes no argument when the hash is empty, so every
   `Table` forwarder (schema_definitions.rb:756, :786, :829, :842, :852,
   :871, :885, :899, :910, :929, :939) records Rails' arity. `CommandRecorder`
   stores arguments verbatim, so the trailing `{}` trails used to append
   showed up in recorded commands. `column` already carried this guard;
   the siblings now do too.

3. `MySQL::SchemaStatements#create_table_definition`
   (mysql/schema_statements.rb:172) was ported twice — as a standalone
   function in `mysql/schema-statements.ts` and again as a live method on
   `AbstractMysqlAdapter`. It now exists once, on `MysqlSchemaStatements`,
   mixed in by `include(AbstractMysqlAdapter, MysqlSchemaStatements)`. The
   `adapter`/`adapterName` strip is gone: `buildCreateTableDefinition` no
   longer injects them, since every implementation supplies `adapter: this`
   the way Rails' `self` does.
