+++ packages/activerecord/src/connection-adapters/abstract/schema-statements-privates.trails.test.ts
@@ -877,7 +877,8 @@ describe("buildCreateTableDefinition routing", () => {
 
   it("extracts _skipValidateOptions into the table definition options", () => {
     const createTableDefinition = vi.fn(
-      (name: string, options: Record<string, unknown>) => new TableDefinition(name, options as any),
+      (name: string, options: Record<string, unknown>) =>
+        new TableDefinition(name, { ...options, adapter: ss } as any),
     );
     const ss = makeStatements({ createTableDefinition });
 
@@ -895,7 +896,7 @@ describe("buildCreateTableDefinition routing", () => {
     const mysql = makeStatements({
       validPrimaryKeyOptions: () => ["limit", "unsigned", "autoIncrement"],
       createTableDefinition: (name: string, options: Record<string, unknown>) =>
-        new MysqlTableDefinition(name, options as any),
+        new MysqlTableDefinition(name, { ...options, adapter: mysql } as any),
     });
 
     expect(
@@ -941,7 +942,7 @@ describe("buildCreateTableDefinition routing", () => {
   it("builds the definition through the adapter's createTableDefinition", () => {
     const marker = Symbol("dialect-td");
     const createTableDefinition = vi.fn((name: string, options: Record<string, unknown>) => {
-      const td: any = new TableDefinition(name, options as any);
+      const td: any = new TableDefinition(name, { ...options, adapter: ss } as any);
       td[marker] = true;
       return td;
     });
